backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
mediapipe
Pillow
numpy
